Add seperate zoom controls module
